requirements.txt
# Web Framework
fastapi==0.115.12
uvicorn[standard]==0.34.0

# Rate Limiting
slowapi==0.1.9

# Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.38

# Crawling
httpx==0.28.1
beautifulsoup4==4.12.3
playwright==1.50.0
lxml==5.3.1

# Cache / Queue
redis==5.2.1
celery==5.4.0

# Utilities
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.7.1
